Suho

+++ Back/src/main/java/com/toogoodtogo/application/shop/ShopService.java
@@ -1,7 +1,10 @@
 package com.toogoodtogo.application.shop;
 
-import com.toogoodtogo.advice.exception.CShopNotFoundException;
-import com.toogoodtogo.advice.exception.CUserNotFoundException;
+import com.toogoodtogo.application.S3Uploader;
+import com.toogoodtogo.application.UploadFileConverter;
+import com.toogoodtogo.domain.security.exceptions.CAccessDeniedException;
+import com.toogoodtogo.domain.shop.exceptions.CShopNotFoundException;
+import com.toogoodtogo.domain.user.exceptions.CUserNotFoundException;
 import com.toogoodtogo.advice.exception.CValidCheckException;
 import com.toogoodtogo.domain.shop.Hours;
 import com.toogoodtogo.domain.shop.Shop;
@@ -15,7 +18,9 @@ import com.toogoodtogo.web.shops.dto.UpdateShopRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -25,6 +30,8 @@ public class ShopService implements ShopUseCase {
     private final ShopRepository shopRepository;
     private final UserRepository userRepository;
     private final ProductRepository productRepository;
+    private final UploadFileConverter uploadFileConverter;
+    private final S3Uploader s3Uploader;
 
     @Override
     @Transactional(readOnly = true)
@@ -46,35 +53,72 @@ public class ShopService implements ShopUseCase {
 
     @Override
     @Transactional
-    public ShopDto addShop(Long managerId, AddShopRequest request) {
-        User manager = userRepository.findById(managerId).orElseThrow(CUserNotFoundException::new); //예외 처리!!
+    public ShopDto addShop(Long managerId, MultipartFile file, AddShopRequest request) throws IOException {
+        User manager = userRepository.findById(managerId).orElseThrow(CUserNotFoundException::new);
         if(shopRepository.findByAddressAndName(request.getAddress(), request.getName()).isPresent())
             throw new CValidCheckException("이미 있는 가게입니다.");
-        Shop shop = Shop.builder()
+
+        // 파일 경로
+        String filePath = uploadFileConverter.parseFileInfo(file, "shopsImage", managerId);
+
+        String fileName;
+        if (filePath.equals("default.png")) fileName = "default.png"; // 기본 이미지
+        else fileName = s3Uploader.upload(file, filePath); // 최종 파일 경로 및 파일 업로드
+
+        Shop newShop = Shop.builder()
                 .user(manager)
                 .name(request.getName())
-                .image(request.getImage())
+                .image(fileName)
                 .category(request.getCategory())
                 .phone(request.getPhone())
                 .address(request.getAddress())
                 .hours(new Hours(request.getOpen(), request.getClose()))
                 .build();
-        return new ShopDto(shopRepository.save(shop));
+        return new ShopDto(shopRepository.save(newShop));
+    }
+
+    private boolean checkAccessOfShop(Long managerId, Long shopId) {
+        return shopRepository.findById(shopId).orElseThrow(CShopNotFoundException::new).getUser().getId().equals(managerId);
     }
 
     @Override
     @Transactional
-    public ShopDto updateShop(Long managerId, Long shopId, UpdateShopRequest request) {
-        Shop modifiedShop = shopRepository.findByUserIdAndId(managerId, shopId).orElseThrow(CShopNotFoundException::new);
-        modifiedShop.update(request.getName(), request.getImage(), request.getCategory(), request.getPhone(), request.getAddress(), new Hours(request.getOpen(), request.getClose()));
+    public ShopDto updateShop(Long managerId, Long shopId, MultipartFile file, UpdateShopRequest request) throws IOException {
+        // 로그인한 유저가 해당 shop 에 대해 권한 가졌는지 체크
+        if(!checkAccessOfShop(managerId, shopId)) throw new CAccessDeniedException();
+        // 수정할 shop
+        Shop modifiedShop = shopRepository.findById(shopId).orElseThrow(CShopNotFoundException::new);
+
+        String filePath;
+        String fileName;
+        if (file.isEmpty()) { // 넘어온 사진이 없으면
+            fileName = modifiedShop.getImage(); // 기존 이미지
+        }
+        else { // 넘어온 사진이 있으면
+            filePath = uploadFileConverter.parseFileInfo(file, "productsImage", modifiedShop.getId());
+            fileName = s3Uploader.updateS3(file, modifiedShop.getImage(), filePath);
+        }
+        
+        // update 리팩토링 필요
+        modifiedShop.update(request.getName(), fileName, request.getCategory(),
+                request.getPhone(), request.getAddress(), new Hours(request.getOpen(), request.getClose()));
         return new ShopDto(modifiedShop);
     }
 
     @Override
     @Transactional
     public String deleteShop(Long managerId, Long shopId) {
-        Shop deleteShop = shopRepository.findByUserIdAndId(managerId, shopId).orElseThrow(CShopNotFoundException::new);
+        // 로그인한 유저가 해당 shop 에 대해 권한 가졌는지 체크
+        if(!checkAccessOfShop(managerId, shopId)) throw new CAccessDeniedException();
+        // 삭제할 shop
+        Shop deleteShop = shopRepository.findById(shopId).orElseThrow(CShopNotFoundException::new);
+
+        // 해당 shop 의 product 들 이미지 삭제
+        s3Uploader.deleteFolderS3("productsImage/" + deleteShop.getId() + "/");
         productRepository.deleteByShopId(deleteShop.getId());
+
+        // 기본 이미지가 아니면 S3에서 이미지 삭제
+        if (!deleteShop.getImage().equals("default.png")) s3Uploader.deleteFileS3(deleteShop.getImage());
         shopRepository.deleteById(deleteShop.getId());
         return "success";
     }
